refactor(nostr): extract DM inbox + conversation cluster into useDmInbox (#703)
Move the NIP-04/NIP-17 DM-inbox + conversation logic out of the
3,565-line NostrProvider into a dedicated `useDmInbox` hook plus two
sibling helper modules, shrinking NostrContext.tsx toward the 1,000-line
cap. Behavior-preserving: no logic, signatures, ordering, or comments
changed — the code was relocated and its provider-owned dependencies
threaded in as hook arguments / returned setters.

- `useDmInbox.ts` (880 lines) owns dmInbox/dmInboxLoading/liveSubArmed/
  amberNip44Permission state, the single-flight + last-refresh + wrap-id
  dedup refs, refreshDmInbox, the logged-out + followPubkeysRef effects,
  armLiveDmSub, and the small DM helper callbacks.
- `nostrLiveDmSub.ts` (516 lines) holds the live kind-4/1059 subscription
  body; the hook's effect calls it and returns its teardown.
- `nostrFetchConversation.ts` (384 lines) holds the thread-open fetch.

The NostrContext context value object is byte-for-byte identical (same
keys, values, dep array). Provider login/logout/switch flows consume the
returned setDmInbox / setAmberNip44Permission / hydrateDmInboxFromCache /
knownWrapIdsRef unchanged. File-size baseline for NostrContext.tsx
lowered 3565 -> 2031.

Gates: tsc clean; jest 700/700 (baseline parity); eslint 0 errors,
warning parity; prettier clean; check-file-size.sh passes.
