+++ src/contexts/NostrContext.tsx
@@ -5,26 +5,17 @@ import React, {
   useEffect,
   useCallback,
   useMemo,
-  useRef,
   startTransition,
 } from 'react';
 import { InteractionManager } from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { File, Paths } from 'expo-file-system';
 import * as SecureStore from 'expo-secure-store';
-import { touchNip17CacheEntry } from '../utils/nip17Cache';
 import * as nip19 from 'nostr-tools/nip19';
 import * as nostrService from '../services/nostrService';
 import * as amberService from '../services/amberService';
 import type { NostrProfile, NostrContact, RelayConfig, SignerType } from '../types/nostr';
 import type { DmInboxEntry } from '../utils/conversationSummaries';
-import {
-  partnerFromRumor,
-  unwrapWrapNsec,
-  unwrapWrapViaNip44,
-  type DecodedRumor,
-} from '../utils/nip17Unwrap';
-import { listPersistedGroupWrapIds } from '../services/groupMessagesStorageService';
 import { getUserRelays, setUserRelays, mergeRelays } from '../services/nostrRelayStorage';
 import { perAccountKey } from '../services/perAccountStorage';
 import {
@@ -43,44 +34,18 @@ import {
   deleteMnemonic,
 } from '../services/walletStorageService';
 import { NSEC_KEY, PUBKEY_KEY, SIGNER_TYPE_KEY } from './nostrAuthKeys';
-import {
-  nip04PlaintextCache,
-  appendLocalDmChains,
-  getMemoisedSecretKey,
-  clearMemoisedSecretKey,
-} from './nostrSecretKeyCache';
-import { notifyDmMessage } from './nostrEventBus';
-import { tryRouteGroupRumor } from './nostrGroupRouting';
-import {
-  yieldToEventLoop,
-  DECRYPT_FRAME_BUDGET_MS,
-  createYieldScheduler,
-  DECRYPT_YIELD_EVERY,
-  NIP17_LOOP_YIELD_EVERY,
-} from './nostrDecryptPacing';
+import { nip04PlaintextCache, clearMemoisedSecretKey } from './nostrSecretKeyCache';
 import {
   AMBER_NIP17_CACHE_KEY_BASE,
   NSEC_NIP17_CACHE_KEY_BASE,
   wrapCacheFileName,
   AMBER_NIP17_ENABLED_KEY_LEGACY,
-  type Nip17CacheEntry,
-  safeParseRecord,
-  writeNip17Cache,
-  DM_INBOX_REFRESH_TTL_MS,
   DM_CONV_CACHE_PREFIX,
   DM_CONV_LAST_SEEN_PREFIX,
-  DM_INBOX_CAP,
-  DM_CONV_CAP,
   inboxCacheKey,
   inboxLastSeenKey,
-  convCacheKey,
-  convLastSeenKey,
-  safeGetDmCacheItem,
-  loadDmInboxFromCache,
-  loadLastSeen,
-  mergeInboxEntries,
-  mergeConversationMessages,
 } from './nostrDmCache';
+import { useDmInbox } from './useDmInbox';
 import {
   CONTACTS_CACHE_KEY_BASE,
   PROFILES_CACHE_KEY_BASE,
@@ -307,38 +272,12 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
   );
   const [signerType, setSignerType] = useState<SignerType | null>(null);
   const [pubkey, setPubkey] = useState<string | null>(null);
-  const [dmInbox, setDmInbox] = useState<DmInboxEntry[]>([]);
-  const [dmInboxLoading, setDmInboxLoading] = useState(false);
-  // Gates the live NIP-17 DM sub useEffect below. False on cold boot
-  // so we don't burn JS-thread cycles unwrapping wraps the user can't
-  // see yet (they're on Home, the Messages tab isn't mounted). Flipped
-  // to true the first time Messages / Conversation / any DM-receiving
-  // surface focuses via `armLiveDmSub()`. Once armed it stays armed for
-  // the rest of the session. Cold-start Home stays responsive because
-  // the per-wrap unwrap/route work moves to after the user has
-  // explicitly chosen to look at messages.
-  const [liveSubArmed, setLiveSubArmed] = useState(false);
-  const [amberNip44Permission, setAmberNip44Permission] = useState<
-    'unknown' | 'granted' | 'denied'
-  >('unknown');
   // Multi-account registry (#288). Mirrors the SecureStore `identities_v1`
   // blob so the drawer header and AccountSwitcherSheet can render without
   // each rendering its own SecureStore round-trip. The `pubkey` state above
   // is the single source of truth for "who is the active identity"; this
   // array is the side-table of all signed-in identities for the switcher.
   const [identities, setIdentities] = useState<StoredIdentity[]>([]);
-
-  // Single-flight guard: coalesce overlapping refreshDmInbox calls (e.g.
-  // useFocusEffect firing while a pull-to-refresh is still in-flight) so
-  // they don't race on the AsyncStorage wrap-id cache.
-  const dmInboxInFlight = useRef<{
-    promise: Promise<void>;
-    includeNonFollows: boolean;
-  } | null>(null);
-  /** `performance.now()` of last successful `refreshDmInbox` completion.
-   * Gate for the `DM_INBOX_REFRESH_TTL_MS` throttle so that Messages-tab
-   * focus doesn't re-fire full relay queries on every tab bounce. */
-  const dmInboxLastRefreshAt = useRef<number>(0);
 
   /**
    * Set of followed pubkeys (lowercase hex). Used to gate who can land a
@@ -416,6 +355,29 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
     const readRelays = relays.filter((r) => r.read).map((r) => r.url);
     return readRelays.length > 0 ? readRelays : nostrService.DEFAULT_RELAYS;
   }, [relays]);
+
+  // DM inbox + conversation cluster (#703). State, refs, callbacks, and
+  // the live-DM subscription effect for the NIP-04/NIP-17 inbox live in
+  // `useDmInbox`. The provider threads in the identity + relay + follow
+  // dependencies these paths close over, and re-exposes the returned
+  // values through the context value below. `setDmInbox`,
+  // `setAmberNip44Permission`, `hydrateDmInboxFromCache`, and
+  // `knownWrapIdsRef` are consumed by the login / logout / switch-identity
+  // flows that own those lifecycle transitions.
+  const {
+    dmInbox,
+    dmInboxLoading,
+    refreshDmInbox,
+    fetchConversation,
+    getCachedConversation,
+    appendLocalDmMessage,
+    armLiveDmSub,
+    amberNip44Permission,
+    hydrateDmInboxFromCache,
+    setDmInbox,
+    setAmberNip44Permission,
+    knownWrapIdsRef,
+  } = useDmInbox({ pubkey, isLoggedIn, signerType, followPubkeys, getReadRelays });
 
   const loadProfile = useCallback(
     async (pk: string, relayUrls: string[], opts?: { force?: boolean }) => {
@@ -568,16 +530,6 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
       console.warn('Failed to load contacts cache:', error);
     }
     return false;
-  }, []);
-
-  /** Eagerly hydrate `dmInbox` from the persisted NIP-17 inbox cache so
-   * the Messages tab paints conversations on cold start instead of
-   * staying blank for the relay-fetch + decrypt loop (~3-5 s). Called
-   * from session-restore + post-login flows; refreshDmInbox handles
-   * its own cache read separately for the delta-fetch path. */
-  const hydrateDmInboxFromCache = useCallback(async (pk: string) => {
-    const cached = await loadDmInboxFromCache(pk);
-    if (cached.length > 0) setDmInbox(cached);
   }, []);
 
   const loadContacts = useCallback(
@@ -1290,7 +1242,14 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
     setIsLoggedIn(false);
 
     nostrService.cleanup();
-  }, [pubkey, wipeAccountCaches, loadContactsFromCache, hydrateDmInboxFromCache]);
+  }, [
+    pubkey,
+    wipeAccountCaches,
+    loadContactsFromCache,
+    hydrateDmInboxFromCache,
+    setAmberNip44Permission,
+    knownWrapIdsRef,
+  ]);
 
   // Flip the active identity to `nextPubkey`. Must already be a
   // registered identity (call `loginWithNsec` / `loginWithAmber`
@@ -1364,7 +1323,16 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
         }
       });
     },
-    [pubkey, loadContactsFromCache, hydrateDmInboxFromCache, loadRelays, loadProfile, loadContacts],
+    [
+      pubkey,
+      loadContactsFromCache,
+      hydrateDmInboxFromCache,
+      loadRelays,
+      loadProfile,
+      loadContacts,
+      setAmberNip44Permission,
+      setDmInbox,
+    ],
   );
 
   // Remove a single identity from the multi-account registry. If
@@ -1939,1508 +1907,6 @@ export const NostrProvider: React.FC<{ children: React.ReactNode }> = ({ childre
     },
     [pubkey, isLoggedIn, signerType, relays],
   );
-
-  /**
-   * Decrypt one NIP-04 payload with whichever signer is active. Returns
-   * null (not throw) on failure so batch callers don't abort the whole
-   * loop on one bad event.
-   */
-  const decryptNip04ViaSigner = useCallback(
-    async (counterpartyPubkey: string, ciphertext: string): Promise<string | null> => {
-      if (!pubkey) return null;
-      try {
-        if (signerType === 'nsec') {
-          const secretKey = await getMemoisedSecretKey(pubkey);
-          if (!secretKey) return null;
-          return await nostrService.decryptNip04WithSecret(
-            secretKey,
-            counterpartyPubkey,
-            ciphertext,
-          );
-        }
-        if (signerType === 'amber') {
-          return await amberService.requestNip04Decrypt(ciphertext, counterpartyPubkey, pubkey);
-        }
-        return null;
-      } catch (error) {
-        if (__DEV__) console.warn('[Nostr] NIP-04 decrypt failed:', error);
-        return null;
-      }
-    },
-    [pubkey, signerType],
-  );
-
-  /**
-   * Silent Amber NIP-44 decrypt wrapped as the callback shape expected by
-   * unwrapWrapViaNip44. Throws on PERMISSION_NOT_GRANTED so the caller
-   * can flip the permission flag and stop iterating rather than falling
-   * back to the Intent dialog (which would flood one dialog per wrap).
-   */
-  const amberNip44DecryptSilent = useCallback(
-    async (ciphertext: string, counterpartyPubkey: string): Promise<string> => {
-      if (!pubkey) throw new Error('Not logged in');
-      return amberService.requestNip44DecryptSilent(ciphertext, counterpartyPubkey, pubkey);
-    },
-    [pubkey],
-  );
-
-  const getCachedConversation = useCallback(
-    async (otherPubkey: string): Promise<ConversationMessage[]> => {
-      if (!pubkey) return [];
-      const normalized = otherPubkey.trim().toLowerCase();
-      if (!/^[0-9a-f]{64}$/.test(normalized)) return [];
-      try {
-        const raw = await safeGetDmCacheItem(convCacheKey(pubkey, normalized));
-        if (!raw) return [];
-        const parsed = JSON.parse(raw);
-        return Array.isArray(parsed) ? parsed : [];
-      } catch {
-        return [];
-      }
-    },
-    [pubkey],
-  );
-
-  const appendLocalDmMessage = useCallback(
-    async (otherPubkey: string, msg: ConversationMessage): Promise<void> => {
-      if (!pubkey) return;
-      const normalized = otherPubkey.trim().toLowerCase();
-      if (!/^[0-9a-f]{64}$/.test(normalized)) return;
-      // Serialize concurrent appends to the same conversation. Without
-      // this, two rapid sends could both read the same `existing` array,
-      // each merge their msg, both write back — last write wins, the
-      // earlier optimistic row is silently lost. Per Copilot review #509.
-      const chainKey = `${pubkey}:${normalized}`;
-      const prev = appendLocalDmChains.get(chainKey) ?? Promise.resolve();
-      const next = prev.then(async () => {
-        try {
-          const key = convCacheKey(pubkey, normalized);
-          const raw = await AsyncStorage.getItem(key);
-          const existing: ConversationMessage[] = raw
-            ? (() => {
-                try {
-                  const parsed = JSON.parse(raw);
-                  return Array.isArray(parsed) ? parsed : [];
-                } catch {
-                  return [];
-                }
-              })()
-            : [];
-          // Dedup on id (same key would arise from a double-tap retry).
-          const map = new Map<string, ConversationMessage>();
-          for (const m of existing) map.set(m.id, m);
-          map.set(msg.id, msg);
-          const merged = Array.from(map.values()).sort((a, b) => a.createdAt - b.createdAt);
-          const capped =
-            merged.length <= DM_CONV_CAP ? merged : merged.slice(merged.length - DM_CONV_CAP);
-          await AsyncStorage.setItem(key, JSON.stringify(capped));
-        } catch {
-          // Swallow — the in-memory setMessages above already painted
-          // the bubble. The remount-after-back regression is precisely
-          // what this method exists to fix, so a write failure is
-          // unfortunate but not destructive (next relay echo will
-          // repopulate the cache).
-        }
-      });
-      // `.catch` on the chain entry so a single failure doesn't poison
-      // every subsequent append on this conversation.
-      appendLocalDmChains.set(
-        chainKey,
-        next.catch(() => {}),
-      );
-      await next;
-    },
-    [pubkey],
-  );
-
-  const fetchConversation = useCallback(
-    async (otherPubkey: string): Promise<ConversationMessage[]> => {
-      if (!pubkey || !isLoggedIn) return [];
-      const normalized = otherPubkey.trim().toLowerCase();
-      if (!/^[0-9a-f]{64}$/.test(normalized)) return [];
-
-      // Perf instrumentation — unconditional (not __DEV__ gated) so we
-      // can grep the same line out of logcat on a production APK to
-      // compare cold-cache vs warm-cache thread opens. Numbers are
-      // counts-only — no plaintext / pubkey logged beyond a short id.
-      const perfStart = performance.now();
-      let nip17CacheHits = 0;
-      let nip17FreshDecrypts = 0;
-      let nip04CacheHits = 0;
-      let nip04FreshDecrypts = 0;
-
-      const readRelays = getReadRelays();
-      const decrypted: ConversationMessage[] = [];
-
-      // PR B: load persisted per-peer conversation + per-peer last-seen.
-      // Merge cached-and-fresh at the end; keep the cache for the next
-      // open so we only ever re-decrypt the (typically 0-few) events
-      // that arrived since last open.
-      const [convRaw, convLastSeen] = await Promise.all([
-        safeGetDmCacheItem(convCacheKey(pubkey, normalized)),
-        loadLastSeen(convLastSeenKey(pubkey, normalized)),
-      ]);
-      const cachedConv: ConversationMessage[] = convRaw
-        ? (() => {
-            try {
-              const parsed = JSON.parse(convRaw);
-              return Array.isArray(parsed) ? parsed : [];
-            } catch {
-              return [];
-            }
-          })()
-        : [];
-
-      // NIP-04 — peer-scoped fetch, two directions, filtered by since.
-      const kind4Events = await nostrService.fetchDirectMessageEvents(
-        pubkey,
-        normalized,
-        readRelays,
-        { since: convLastSeen },
-      );
-      // Two-pass decrypt with module-level LRU cache:
-      //  1. Pull plaintext synchronously for events already in the
-      //     cache — no decrypt round-trip, no Amber IPC, no CPU.
-      //  2. Decrypt the misses in parallel (`Promise.all`). For nsec
-      //     this drains the JS queue faster than a serial for-await
-      //     loop; for Amber the IPC round-trips pipeline. Chunk the
-      //     Promise.all in slices of DECRYPT_YIELD_EVERY to yield to
-      //     the UI thread between batches on very long threads.
-      const freshDecryptTargets: {
-        idx: number;
-        counterparty: string;
-        ev: (typeof kind4Events)[0];
-      }[] = [];
-      const cachedPlaintexts: {
-        idx: number;
-        fromMe: boolean;
-        text: string;
-        ev: (typeof kind4Events)[0];
-      }[] = [];
-      for (let i = 0; i < kind4Events.length; i++) {
-        const ev = kind4Events[i];
-        const fromMe = ev.pubkey === pubkey;
-        const counterparty = fromMe ? normalized : ev.pubkey.toLowerCase();
-        const hit = nip04PlaintextCache.get(ev.id);
-        if (hit !== undefined) {
-          nip04CacheHits++;
-          cachedPlaintexts.push({ idx: i, fromMe, text: hit, ev });
-        } else {
-          freshDecryptTargets.push({ idx: i, counterparty, ev });
-        }
-      }
-      // Parallel decrypt of misses, in yield-able chunks.
-      const freshResults: ({
-        idx: number;
-        fromMe: boolean;
-        text: string;
-        ev: (typeof kind4Events)[0];
-      } | null)[] = [];
-      for (let i = 0; i < freshDecryptTargets.length; i += DECRYPT_YIELD_EVERY) {
-        const batch = freshDecryptTargets.slice(i, i + DECRYPT_YIELD_EVERY);
-        const batchResults = await Promise.all(
-          batch.map(async (t) => {
-            nip04FreshDecrypts++;
-            const plaintext = await decryptNip04ViaSigner(t.counterparty, t.ev.content);
-            if (plaintext === null) return null;
-            // Cache the successful decrypt. Event ids are immutable so
-            // we can store unconditionally — no staleness possible.
-            nip04PlaintextCache.set(t.ev.id, plaintext);
-            const fromMe = t.ev.pubkey === pubkey;
-            return { idx: t.idx, fromMe, text: plaintext, ev: t.ev };
-          }),
-        );
-        freshResults.push(...batchResults);
-        if (i + DECRYPT_YIELD_EVERY < freshDecryptTargets.length) await yieldToEventLoop();
-      }
-      // Merge cached + fresh preserving original event order.
-      const orderedByIndex = new Array<ConversationMessage | null>(kind4Events.length).fill(null);
-      for (const c of cachedPlaintexts) {
-        orderedByIndex[c.idx] = {
-          id: c.ev.id,
-          fromMe: c.fromMe,
-          text: c.text,
-          createdAt: c.ev.created_at,
-        };
-      }
-      for (const r of freshResults) {
-        if (!r) continue;
-        orderedByIndex[r.idx] = {
-          id: r.ev.id,
-          fromMe: r.fromMe,
-          text: r.text,
-          createdAt: r.ev.created_at,
-        };
-      }
-      for (const m of orderedByIndex) if (m !== null) decrypted.push(m);
-
-      // NIP-17 — partner pubkey is hidden inside the encrypted rumor,
-      // so we can't peer-scope at the relay. `refreshDmInbox` (which
-      // the Messages tab fires on focus with a 30s TTL) already
-      // decrypts every wrap addressed to us and writes the plaintext
-      // keyed by wrap id to the persistent cache. Serve the NIP-17
-      // portion of THIS thread from that cache first — if the cache
-      // has ANY entries, we skip the expensive inbox-wide relay
-      // fetch entirely (#190).
-      //
-      // Cold-cache fallback: if the cache has no entries at all (first
-      // app run post-login, or just-logged-out-logged-back-in) we
-      // still hit the relay so the thread renders even before any
-      // refreshDmInbox has fired. Subsequent opens short-circuit.
-      //
-      // Staleness tradeoff: a wrap that arrived in the last <30s and
-      // hasn't been pulled by refreshDmInbox yet won't show until the
-      // next tab focus. For a chat UX that's a non-issue.
-      const signerWrapCacheKey =
-        signerType === 'nsec'
-          ? perAccountKey(NSEC_NIP17_CACHE_KEY_BASE, pubkey)
-          : signerType === 'amber'
-            ? perAccountKey(AMBER_NIP17_CACHE_KEY_BASE, pubkey)
-            : null;
-      const wrapCacheRaw = signerWrapCacheKey ? await safeGetDmCacheItem(signerWrapCacheKey) : null;
-      const wrapCache = safeParseRecord<Nip17CacheEntry>(wrapCacheRaw);
-      const cachedWrapEntries = Object.values(wrapCache);
-      let skippedInboxFetch = false;
-      let fastPathTouched = 0;
-      if (cachedWrapEntries.length > 0) {
-        // Cache populated — serve peer-matching wraps directly, skip relay fetch.
-        for (const entry of cachedWrapEntries) {
-          nip17CacheHits++;
-          if (entry.partnerPubkey !== normalized) continue;
-          // LRU touch (#193) — opening this thread is a "use" of these entries; without the touch they age out FIFO and a thread the user re-opens regularly can be evicted just because newer wraps arrived first.
-          touchNip17CacheEntry(wrapCache, entry.wrapId);
-          fastPathTouched++;
-          decrypted.push({
-            id: entry.wrapId,
-            fromMe: entry.fromMe,
-            text: entry.text,
-            createdAt: entry.createdAt,
-          });
-        }
-        skippedInboxFetch = true;
-      }
-      // Persist the touched-cache so LRU order survives restarts.
-      if (fastPathTouched > 0 && signerWrapCacheKey) {
-        await writeNip17Cache(signerWrapCacheKey, wrapCache);
-      }
-      const inboxLastSeenForWraps = skippedInboxFetch
-        ? undefined
-        : await loadLastSeen(inboxLastSeenKey(pubkey));
-      const { kind1059 } = skippedInboxFetch
-        ? {
-            kind1059: [] as Awaited<ReturnType<typeof nostrService.fetchInboxDmEvents>>['kind1059'],
-          }
-        : await nostrService.fetchInboxDmEvents(pubkey, readRelays, {
-            since: inboxLastSeenForWraps,
-          });
-      if (kind1059.length > 0) {
-        const onSkip = (reason: string, wrapId: string) => {
-          if (__DEV__) console.warn(`[Nostr] NIP-17 thread unwrap skip (${wrapId}): ${reason}`);
-        };
-        if (signerType === 'nsec') {
-          const secretKey = await getMemoisedSecretKey(pubkey);
-          if (secretKey) {
-            // Reuse the persistent wrap-id cache populated by
-            // refreshDmInbox (#176). For wraps that aren't cached yet
-            // (typically arrived between the last inbox refresh and
-            // this thread open), we decrypt AND write them back so the
-            // next thread open across ANY conversation can short-circuit
-            // without waiting for the next inbox refresh.
-            const nsecCacheKey = perAccountKey(NSEC_NIP17_CACHE_KEY_BASE, pubkey);
-            const raw = await safeGetDmCacheItem(nsecCacheKey);
-            const cache = safeParseRecord<Nip17CacheEntry>(raw);
-            const newlyCached: Nip17CacheEntry[] = [];
-            let nip17Decrypted = 0;
-            let threadTouched = 0;
-            for (const wrap of kind1059) {
-              const cached = cache[wrap.id];
-              if (cached) {
-                nip17CacheHits++;
-                if (cached.partnerPubkey !== normalized) continue;
-                // LRU touch (#193) — see fast-path above for rationale.
-                touchNip17CacheEntry(cache, wrap.id);
-                threadTouched++;
-                decrypted.push({
-                  id: wrap.id,
-                  fromMe: cached.fromMe,
-                  text: cached.text,
-                  createdAt: cached.createdAt,
-                });
-                continue;
-              }
-              nip17FreshDecrypts++;
-              const rumor = unwrapWrapNsec(wrap, secretKey, onSkip);
-              if (++nip17Decrypted % DECRYPT_YIELD_EVERY === 0) await yieldToEventLoop();
-              if (!rumor) continue;
-              // If this is a multi-recipient group rumor, route it to
-              // the group store and skip 1:1 caching. Opening a DM
-              // thread shouldn't backfill group rumors into the 1:1
-              // cache — they belong to GroupConversationScreen.
-              const routeResult = await tryRouteGroupRumor(rumor, pubkey, wrap.id);
-              if (routeResult.kind !== 'not-group') continue;
-              const partnership = partnerFromRumor(rumor, pubkey);
-              if (!partnership) continue;
-              // Cache every successfully decrypted wrap, even if it
-              // belongs to a different thread — cache is keyed by wrap
-              // id, not by thread, so later opens of OTHER threads
-              // benefit too. Filter to this thread's partner only for
-              // the render-side `decrypted` array.
-              const entry: Nip17CacheEntry = {
-                id: wrap.id,
-                wrapId: wrap.id,
-                partnerPubkey: partnership.partnerPubkey,
-                fromMe: partnership.fromMe,
-                createdAt: rumor.created_at,
-                text: rumor.content,
-                wireKind: rumor.kind,
-              };
-              cache[wrap.id] = entry;
-              newlyCached.push(entry);
-              if (partnership.partnerPubkey !== normalized) continue;
-              decrypted.push({
-                id: wrap.id,
-                fromMe: partnership.fromMe,
-                text: rumor.content,
-                createdAt: rumor.created_at,
-              });
-            }
-            if (newlyCached.length > 0 || threadTouched > 0) {
-              await writeNip17Cache(nsecCacheKey, cache);
-            }
-          }
-        } else if (signerType === 'amber') {
-          const amberCacheKey = perAccountKey(AMBER_NIP17_CACHE_KEY_BASE, pubkey);
-          const raw = await safeGetDmCacheItem(amberCacheKey);
-          const cache = safeParseRecord<Nip17CacheEntry>(raw);
-          let threadTouched = 0;
-          for (const wrap of kind1059) {
-            const cached = cache[wrap.id];
-            if (cached) {
-              nip17CacheHits++;
-              if (cached.partnerPubkey !== normalized) continue;
-              touchNip17CacheEntry(cache, wrap.id);
-              threadTouched++;
-              decrypted.push({
-                id: wrap.id,
-                fromMe: cached.fromMe,
-                text: cached.text,
-                createdAt: cached.createdAt,
-              });
-              continue;
-            }
-            nip17FreshDecrypts++;
-            // Thread view falls back to the Intent dialog if the silent path rejects — the user has actively opened this thread, one approval prompt per wrap is fine. Inbox refresh uses the silent-only path to avoid the flood; cached entries cover the hot path.
-            try {
-              const rumor = await unwrapWrapViaNip44(
-                wrap,
-                (ct, cp) => amberService.requestNip44Decrypt(ct, cp, pubkey),
-                onSkip,
-              );
-              if (!rumor) continue;
-              const routeResult = await tryRouteGroupRumor(rumor, pubkey, wrap.id);
-              if (routeResult.kind !== 'not-group') continue;
-              const partnership = partnerFromRumor(rumor, pubkey);
-              if (!partnership || partnership.partnerPubkey !== normalized) continue;
-              decrypted.push({
-                id: wrap.id,
-                fromMe: partnership.fromMe,
-                text: rumor.content,
-                createdAt: rumor.created_at,
-              });
-            } catch (error) {
-              if (__DEV__) console.warn('[Nostr] Amber NIP-17 thread unwrap failed:', error);
-            }
-          }
-          if (threadTouched > 0) {
-            await writeNip17Cache(amberCacheKey, cache);
-          }
-        }
-      }
-
-      // PR B: merge fresh decrypt results with what we had cached
-      // from previous opens. Fresh takes precedence via `mergeConversationMessages`
-      // Map semantics so re-ordered or edited events (rare) land right.
-      const merged = mergeConversationMessages(cachedConv, decrypted, DM_CONV_CAP);
-
-      // Single-line perf summary — grep `[Perf] fetchConversation` out
-      // of logcat to compare cold-cache vs warm-cache thread opens.
-      // Cold cache shows `hits=0, fresh=N` — whole inbox decrypted.
-      // Warm cache shows `hits≈N, fresh=0` — all cache short-circuits.
-      console.log(
-        `[Perf] fetchConversation(${normalized.slice(0, 8)}): ` +
-          `${(performance.now() - perfStart).toFixed(0)}ms, ` +
-          `k4=${kind4Events.length} (hits=${nip04CacheHits}, fresh=${nip04FreshDecrypts}), ` +
-          `k1059=${nip17CacheHits + nip17FreshDecrypts} (hits=${nip17CacheHits}, fresh=${nip17FreshDecrypts}, skippedFetch=${skippedInboxFetch}), ` +
-          `since=${convLastSeen ?? 0}, ` +
-          `cached=${cachedConv.length}, ` +
-          `merged=${merged.length}`,
-      );
-
-      // Persist merged list + new per-peer last-seen so next open of
-      // THIS thread sees only the delta. Fire-and-forget; the caller
-      // gets its data immediately via `merged`. kind-1059 deliberately
-      // excluded — wrap timestamps are randomized per NIP-59 and would
-      // poison the kind-4 since cursor (same reasoning as the inbox
-      // path; see fetchInboxDmEvents + refreshDmInbox).
-      const newConvLastSeen = Math.max(convLastSeen ?? 0, ...kind4Events.map((e) => e.created_at));
-      Promise.all([
-        AsyncStorage.setItem(convCacheKey(pubkey, normalized), JSON.stringify(merged)).catch(
-          () => {},
-        ),
-        newConvLastSeen > (convLastSeen ?? 0)
-          ? AsyncStorage.setItem(
-              convLastSeenKey(pubkey, normalized),
-              String(newConvLastSeen),
-            ).catch(() => {})
-          : Promise.resolve(),
-      ]).catch(() => {});
-
-      return merged;
-    },
-    [pubkey, isLoggedIn, signerType, getReadRelays, decryptNip04ViaSigner],
-  );
-
-  const refreshDmInbox = useCallback(
-    async (opts?: RefreshDmInboxOptions): Promise<void> => {
-      if (!pubkey || !isLoggedIn) {
-        setDmInbox([]);
-        return;
-      }
-      // [PerfBlock] timing bracket — surfaces the wall-clock cost of
-      // a full inbox refresh including NIP-17 decrypt loops. Look for
-      // matched `refreshDmInbox: …ms` pairs in logcat to isolate
-      // multi-second freezes that coincide with this call. #554.
-      const __perfBlockStart = performance.now();
-      const signal = opts?.signal;
-      // Dev-only "Following only=off" bypass — read once at the top so
-      // the closure captures a stable value across the async work below.
-      // When true, all six follow-gate `continue`s in the decrypt loops
-      // become no-ops AND the cache hydrate skips its filter so the
-      // already-cached unfollowed entries don't get masked.
-      const includeNonFollows = opts?.includeNonFollows === true;
-      // Freshness TTL: skip the refresh entirely if the previous one
-      // finished within DM_INBOX_REFRESH_TTL_MS, unless the caller
-      // explicitly opts into a forced refresh (pull-to-refresh). The
-      // Messages tab's `useFocusEffect` uses the default TTL path so
-      // tab-bouncing doesn't retrigger expensive relay+decrypt work.
-      if (!opts?.force) {
-        const age = performance.now() - dmInboxLastRefreshAt.current;
-        if (dmInboxLastRefreshAt.current > 0 && age < DM_INBOX_REFRESH_TTL_MS) {
-          return;
-        }
-      }
-      // Single-flight: piggy-back on in-flight task ONLY when its includeNonFollows matches; otherwise wait then re-run with the wider option.
-      if (dmInboxInFlight.current) {
-        if (dmInboxInFlight.current.includeNonFollows === includeNonFollows) {
-          return dmInboxInFlight.current.promise;
-        }
-        await dmInboxInFlight.current.promise;
-      }
-
-      // Capture local references once so the closure isn't affected by
-      // mid-flight signer / identity changes. If we detect pubkey/signerType
-      // has changed by the time we're about to commit, we bail without
-      // mutating state to avoid leaking entries into the wrong session.
-      const refreshForPubkey = pubkey;
-      // Local helper: encapsulates the follow gate so all seven sites in
-      // the cache hydrate + NIP-04 + NIP-17 decrypt loops + final merge
-      // reuse the same predicate. When includeNonFollows is true the
-      // gate is a no-op (every pubkey passes), so callers can opt out
-      // of the parental-control filter from a single switch.
-      const refreshForSigner = signerType;
-      const refreshFollows = followPubkeys;
-      const passesFollowGate = (pk: string): boolean => includeNonFollows || refreshFollows.has(pk);
-
-      const task = (async () => {
-        setDmInboxLoading(true);
-        try {
-          const readRelays = getReadRelays();
-          const refreshStart = performance.now();
-          let nip04CacheHits = 0;
-          let nip04FreshDecrypts = 0;
-          // NIP-17 perf counters — emitted in the `[Perf] refreshDmInbox`
-          // line so #193 can be tracked post-merge. `nip17Hits` /
-          // `nip17Misses` capture the cache-hit ratio per refresh;
-          // `nip17Evictions` shows whether the 5000-cap is actually
-          // squeezing entries out (was previously invisible — the FIFO
-          // sort-and-slice path ran silently).
-          let nip17Hits = 0;
-          let nip17Misses = 0;
-          let nip17Evictions = 0;
-          let nip17CacheSize = 0;
-          // Number of actual `setTimeout(0)` yields the NIP-17 loop
-          // performed this refresh — emitted in the [Perf] nip17-cache
-          // line so we can track how often the new frame-budget
-          // scheduler trips (#532). Higher = more breathing room
-          // given back to the UI thread.
-          let nip17YieldCount = 0;
-
-          // PR B: load persisted inbox + last-seen so we can (a) paint
-          // cached entries before the relay round-trip finishes and
-          // (b) only fetch events newer than the last one we saw.
-          const [cachedInboxRaw, lastSeen] = await Promise.all([
-            safeGetDmCacheItem(inboxCacheKey(refreshForPubkey)),
-            loadLastSeen(inboxLastSeenKey(refreshForPubkey)),
-          ]);
-          const cachedInbox: DmInboxEntry[] = cachedInboxRaw
-            ? (() => {
-                try {
-                  const parsed = JSON.parse(cachedInboxRaw);
-                  return Array.isArray(parsed) ? parsed : [];
-                } catch {
-                  return [];
-                }
-              })()
-            : [];
-          // Render the cached entries immediately so the Messages tab
-          // isn't blank while the relay fetches the delta. The followers
-          // set may have changed since the cache was written; re-apply
-          // the filter here so unfollowed senders don't resurrect.
-          if (cachedInbox.length > 0) {
-            const filteredCache = cachedInbox.filter((e) => passesFollowGate(e.partnerPubkey));
-            setDmInbox(filteredCache);
-          }
-
-          // For pull-to-refresh / force refresh, skip the `since` filter
-          // entirely. NIP-59 wraps have a randomised `created_at` (up to
-          // 2 days back), so a `since` cutoff is unreliable for catching
-          // freshly-published wraps — the relay will drop wraps whose
-          // randomised stamp falls behind the cutoff. The wrap-id cache
-          // dedupes the re-fetched bytes, so the cost of dropping the
-          // floor is just the relay round-trip, not re-decrypt. Group
-          // messages especially benefit since GroupsScreen / GroupConv
-          // open with `force: true` to chase newly-arrived rumors.
-          const { kind4, kind1059 } = await nostrService.fetchInboxDmEvents(
-            refreshForPubkey,
-            readRelays,
-            opts?.force ? {} : { since: lastSeen },
-          );
-          if (signal?.aborted) return;
-          const entries: DmInboxEntry[] = [];
-
-          // NIP-04 — partner pubkey is in the envelope, so we can apply
-          // the follow filter BEFORE decrypting. A non-followed sender
-          // never gets a round-trip through Amber, let alone land in
-          // state. Same cache/parallel pattern as fetchConversation:
-          // pull cached plaintext synchronously, decrypt misses in
-          // DECRYPT_YIELD_EVERY-sized parallel batches.
-          const k4Targets: {
-            ev: (typeof kind4)[0];
-            fromMe: boolean;
-            partnerPubkey: string;
-          }[] = [];
-          for (const ev of kind4) {
-            const fromMe = ev.pubkey.toLowerCase() === refreshForPubkey;
-            const partnerPubkey = (
-              fromMe ? (ev.tags.find((t) => t[0] === 'p')?.[1] ?? '') : ev.pubkey
-            ).toLowerCase();
-            if (!/^[0-9a-f]{64}$/.test(partnerPubkey)) continue;
-            if (!passesFollowGate(partnerPubkey)) continue;
-            k4Targets.push({ ev, fromMe, partnerPubkey });
-          }
-          // Fast pass — cache lookup only.
-          const k4Misses: typeof k4Targets = [];
-          for (const t of k4Targets) {
-            const hit = nip04PlaintextCache.get(t.ev.id);
-            if (hit !== undefined) {
-              nip04CacheHits++;
-              entries.push({
-                id: t.ev.id,
-                partnerPubkey: t.partnerPubkey,
-                fromMe: t.fromMe,
-                createdAt: t.ev.created_at,
-                text: hit,
-                wireKind: 4,
-              });
-            } else {
-              k4Misses.push(t);
-            }
-          }
-          // Slow pass — parallel decrypt of misses in yield-able chunks.
-          for (let i = 0; i < k4Misses.length; i += DECRYPT_YIELD_EVERY) {
-            if (signal?.aborted) return;
-            const batch = k4Misses.slice(i, i + DECRYPT_YIELD_EVERY);
-            const batchResults = await Promise.all(
-              batch.map(async (t) => {
-                nip04FreshDecrypts++;
-                const plaintext = await decryptNip04ViaSigner(t.partnerPubkey, t.ev.content);
-                if (plaintext === null) return null;
-                nip04PlaintextCache.set(t.ev.id, plaintext);
-                return { t, plaintext };
-              }),
-            );
-            for (const r of batchResults) {
-              if (!r) continue;
-              entries.push({
-                id: r.t.ev.id,
-                partnerPubkey: r.t.partnerPubkey,
-                fromMe: r.t.fromMe,
-                createdAt: r.t.ev.created_at,
-                text: r.plaintext,
-                wireKind: 4,
-              });
-            }
-            if (i + DECRYPT_YIELD_EVERY < k4Misses.length) await yieldToEventLoop();
-          }
-          if (signal?.aborted) return;
-
-          // NIP-17 — partner pubkey is INSIDE the encrypted rumor, so we
-          // have to decrypt to know who sent it. For the nsec signer this
-          // is cheap pure-JS, so iterate freely and drop non-follows after
-          // decrypt. For Amber, guard behind the opt-in toggle + silent-only
-          // decrypt path: if Amber hasn't pre-approved nip44_decrypt, we
-          // flip amberNip44Permission='denied' so Account can prompt the
-          // user for a one-time grant, and stop iterating instead of
-          // flooding dialogs.
-          const onSkip = (reason: string, wrapId: string) => {
-            if (__DEV__) console.warn(`[Nostr] NIP-17 inbox unwrap skip (${wrapId}): ${reason}`);
-          };
-
-          if (refreshForSigner === 'nsec' && kind1059.length > 0) {
-            const secretKey = await getMemoisedSecretKey(refreshForPubkey);
-            if (secretKey) {
-              // Persistent wrap-id cache mirroring the Amber branch. Only
-              // ever contains rumors from followed senders — see the
-              // filter gate below. This is the fix for #176. Per-account
-              // namespaced (#288).
-              const nsecCacheKey = perAccountKey(NSEC_NIP17_CACHE_KEY_BASE, refreshForPubkey);
-              const raw = await safeGetDmCacheItem(nsecCacheKey);
-              const cache = safeParseRecord<Nip17CacheEntry>(raw);
-              const newlyCached: Nip17CacheEntry[] = [];
-              let unfollowedPurged = 0;
-              let touched = 0;
-              // Frame-budget scheduler (#532): yield whenever we've held
-              // the JS thread for >= DECRYPT_FRAME_BUDGET_MS, with the
-              // count-based modulo kept as a safety cap. On abort, the
-              // scheduler hard-cancels any pending setTimeout so the
-              // loop unwinds in the next microtask instead of waiting
-              // out one more scheduler round-trip.
-              const nsecYield = createYieldScheduler({
-                signal,
-                safetyEvery: NIP17_LOOP_YIELD_EVERY,
-              });
-              try {
-                for (const wrap of kind1059) {
-                  // Time-budget yield + abort check (#286, #532). Covers
-                  // the cache-hit path too — without it, a long run of
-                  // cache hits walks the whole kind1059 list synchronously
-                  // and any back-tap during refresh appears frozen.
-                  await nsecYield.maybeYield();
-                  if (signal?.aborted) return;
-                  const cached = cache[wrap.id];
-                  if (cached) {
-                    nip17Hits++;
-                    // Cache entry exists → it was from a followed sender
-                    // when first stored. Re-check against the *current*
-                    // follow set so unfollowed partners don't keep
-                    // surfacing from cache. Purge the stale entry so we
-                    // don't keep dragging it through every refresh until
-                    // the 5000-cap LRU finally evicts it.
-                    if (!passesFollowGate(cached.partnerPubkey)) {
-                      delete cache[wrap.id];
-                      unfollowedPurged++;
-                      continue;
-                    }
-                    // LRU touch (#193): re-insert at the tail so this hot
-                    // entry survives the next overflow eviction. Without
-                    // this the cache is FIFO-by-first-write — a thread
-                    // the user re-opens regularly can be evicted just
-                    // because 5000 newer wraps happened to arrive first.
-                    touchNip17CacheEntry(cache, wrap.id);
-                    touched++;
-                    entries.push({
-                      id: cached.wrapId,
-                      partnerPubkey: cached.partnerPubkey,
-                      fromMe: cached.fromMe,
-                      createdAt: cached.createdAt,
-                      text: cached.text,
-                      wireKind: cached.wireKind,
-                    });
-                    continue;
-                  }
-                  nip17Misses++;
-                  const rumor = unwrapWrapNsec(wrap, secretKey, onSkip);
-                  // No per-decrypt yield here: the frame-budget scheduler
-                  // at the top of the loop already yields whenever the
-                  // accumulated work exceeds DECRYPT_FRAME_BUDGET_MS,
-                  // which captures the cost of unwrapWrapNsec naturally.
-                  if (!rumor) continue;
-                  // Multi-recipient (group) rumors: route to group storage
-                  // and short-circuit the DM-inbox path. The 1:1 inbox
-                  // never sees group messages — they belong to a different
-                  // surface (GroupConversationScreen).
-                  const routeResult = await tryRouteGroupRumor(rumor, refreshForPubkey, wrap.id);
-                  if (routeResult.kind !== 'not-group') continue;
-                  const partnership = partnerFromRumor(rumor, refreshForPubkey);
-                  if (!partnership) continue;
-                  // B1 — drop non-follows at the data layer. No caching, no
-                  // state. The filter is load-bearing ("parental control"),
-                  // so it runs here not in the view.
-                  if (!passesFollowGate(partnership.partnerPubkey)) continue;
-                  const entry: Nip17CacheEntry = {
-                    id: wrap.id,
-                    wrapId: wrap.id,
-                    partnerPubkey: partnership.partnerPubkey,
-                    fromMe: partnership.fromMe,
-                    createdAt: rumor.created_at,
-                    text: rumor.content,
-                    wireKind: rumor.kind,
-                  };
-                  cache[wrap.id] = entry;
-                  newlyCached.push(entry);
-                  entries.push({
-                    id: entry.id,
-                    partnerPubkey: entry.partnerPubkey,
-                    fromMe: entry.fromMe,
-                    createdAt: entry.createdAt,
-                    text: entry.text,
-                    wireKind: entry.wireKind,
-                  });
-                }
-              } finally {
-                nsecYield.dispose();
-              }
-              nip17YieldCount += nsecYield.yieldCount;
-
-              // Persist if we mutated the cache for any reason: new
-              // entries, follow-set purges, or LRU touches (#193) — the
-              // touch reorders insertion order, and we need that order
-              // on disk for it to survive app restart.
-              if (newlyCached.length > 0 || unfollowedPurged > 0 || touched > 0) {
-                nip17Evictions += await writeNip17Cache(nsecCacheKey, cache);
-              }
-              nip17CacheSize = Object.keys(cache).length;
-            }
-          } else if (refreshForSigner === 'amber' && kind1059.length > 0) {
-            // Always run the unwrap loop — Amber's silent content-resolver path returns PERMISSION_NOT_GRANTED on the first wrap if the user hasn't granted nip44_decrypt yet, which we surface via setAmberNip44Permission('denied') so NostrScreen can show the one-shot "Grant permission in Amber" button. Closes #404.
-            // Persistent cache keyed by wrap id. Only ever contains rumors from *followed* senders — see the filter gate below. Per-account namespaced (#288).
-            const amberCacheKey = perAccountKey(AMBER_NIP17_CACHE_KEY_BASE, refreshForPubkey);
-            const raw = await safeGetDmCacheItem(amberCacheKey);
-            const cache = safeParseRecord<Nip17CacheEntry>(raw);
-            const newlyCached: Nip17CacheEntry[] = [];
-            let permissionDenied = false;
-            let touched = 0;
-            let unfollowedPurged = 0;
-            // Frame-budget scheduler (#532) — see nsec branch above.
-            const amberYield = createYieldScheduler({
-              signal,
-              safetyEvery: NIP17_LOOP_YIELD_EVERY,
-            });
-            try {
-              for (const wrap of kind1059) {
-                // Time-budget yield + abort check (#286, #532) — see nsec branch above for rationale.
-                await amberYield.maybeYield();
-                if (signal?.aborted) return;
-                const cached = cache[wrap.id];
-                if (cached) {
-                  nip17Hits++;
-                  // Re-check against current follow set; purge if no longer followed so we don't drag stale entries through every refresh.
-                  if (!passesFollowGate(cached.partnerPubkey)) {
-                    delete cache[wrap.id];
-                    unfollowedPurged++;
-                    continue;
-                  }
-                  touchNip17CacheEntry(cache, wrap.id);
-                  touched++;
-                  entries.push({
-                    id: cached.wrapId,
-                    partnerPubkey: cached.partnerPubkey,
-                    fromMe: cached.fromMe,
-                    createdAt: cached.createdAt,
-                    text: cached.text,
-                    wireKind: cached.wireKind,
-                  });
-                  continue;
-                }
-                nip17Misses++;
-                // Uncached — unwrap via Amber's silent content-resolver path.
-                // If Amber hasn't granted blanket nip44_decrypt permission,
-                // this throws PERMISSION_NOT_GRANTED and we stop iterating.
-                try {
-                  const rumor = await unwrapWrapViaNip44(wrap, amberNip44DecryptSilent, onSkip);
-                  if (!rumor) continue;
-                  // Multi-recipient (group) rumors: route to group storage
-                  // and short-circuit the DM-inbox path.
-                  const routeResult = await tryRouteGroupRumor(rumor, refreshForPubkey, wrap.id);
-                  if (routeResult.kind !== 'not-group') continue;
-                  const partnership = partnerFromRumor(rumor, refreshForPubkey);
-                  if (!partnership) continue;
-                  // B1 — never cache rumors from non-followed senders. The
-                  // cost is re-decrypting them on the next refresh, but the
-                  // silent path is ~1 ms per call and keeps plaintext off
-                  // AsyncStorage.
-                  if (!passesFollowGate(partnership.partnerPubkey)) continue;
-                  const entry: Nip17CacheEntry = {
-                    id: wrap.id,
-                    wrapId: wrap.id,
-                    partnerPubkey: partnership.partnerPubkey,
-                    fromMe: partnership.fromMe,
-                    createdAt: rumor.created_at,
-                    text: rumor.content,
-                    wireKind: rumor.kind,
-                  };
-                  cache[wrap.id] = entry;
-                  newlyCached.push(entry);
-                  entries.push({
-                    id: entry.id,
-                    partnerPubkey: entry.partnerPubkey,
-                    fromMe: entry.fromMe,
-                    createdAt: entry.createdAt,
-                    text: entry.text,
-                    wireKind: entry.wireKind,
-                  });
-                } catch (error) {
-                  const code = (error as { code?: string })?.code;
-                  const message = (error as Error)?.message ?? '';
-                  if (code === 'PERMISSION_NOT_GRANTED' || /PERMISSION_NOT_GRANTED/.test(message)) {
-                    permissionDenied = true;
-                    if (__DEV__) {
-                      console.log(
-                        `[Nostr] Amber NIP-44 permission not granted — stopping NIP-17 unwrap for this refresh`,
-                      );
-                    }
-                    break;
-                  }
-                  if (__DEV__) console.warn('[Nostr] Amber NIP-17 unwrap failed:', error);
-                }
-              }
-            } finally {
-              amberYield.dispose();
-            }
-            nip17YieldCount += amberYield.yieldCount;
-
-            setAmberNip44Permission(permissionDenied ? 'denied' : 'granted');
-
-            // Persist on new entries, LRU touches (#193 — touches reorder insertion order which we need on disk), or unfollowed-partner purges (so the purge survives the next launch).
-            if (newlyCached.length > 0 || touched > 0 || unfollowedPurged > 0) {
-              nip17Evictions += await writeNip17Cache(amberCacheKey, cache);
-            }
-            nip17CacheSize = Object.keys(cache).length;
-          }
-
-          // Identity-change guard: if the user logged out or switched signer
-          // while we were mid-flight, don't leak these entries into a
-          // different session's state. Abort signal is treated the same way:
-          // if the navigating-away screen has signalled cancel, skip the
-          // commit so we don't pay the merge / persist cost.
-          if (refreshForPubkey !== pubkey || refreshForSigner !== signerType) return;
-          if (signal?.aborted) return;
-
-          // PR B: merge cached-with-fresh, keep at most DM_INBOX_CAP
-          // entries (newest-first), then persist + update last-seen.
-          const merged = mergeInboxEntries(cachedInbox, entries, DM_INBOX_CAP);
-          const filteredFinal = merged.filter((e) => passesFollowGate(e.partnerPubkey));
-
-          // Perf summary: one line per refresh, grep with `\[Perf\] refreshDmInbox`.
-          // The `nip17-cache` segment (#193) lets us see at a glance
-          // whether the LRU swap is keeping hot entries warm: a healthy
-          // long-running inbox should converge to hits >> misses with
-          // size pinned at the 5000 cap and a non-zero evictions counter
-          // each refresh once the cap is reached.
-          console.log(
-            `[Perf] refreshDmInbox: ` +
-              `${(performance.now() - refreshStart).toFixed(0)}ms, ` +
-              `k4=${kind4.length} (hits=${nip04CacheHits}, fresh=${nip04FreshDecrypts}), ` +
-              `k1059=${kind1059.length}, ` +
-              `since=${lastSeen ?? 0}, ` +
-              `fresh=${entries.length}, ` +
-              `merged=${merged.length}, ` +
-              `rendered=${filteredFinal.length}`,
-          );
-          console.log(
-            `[Perf] nip17-cache: ` +
-              `hits=${nip17Hits}, ` +
-              `misses=${nip17Misses}, ` +
-              `evictions=${nip17Evictions}, ` +
-              `size=${nip17CacheSize}, ` +
-              `yields=${nip17YieldCount} (budget=${DECRYPT_FRAME_BUDGET_MS}ms, cap=${NIP17_LOOP_YIELD_EVERY})`,
-          );
-
-          setDmInbox(filteredFinal);
-
-          // Persist merged list + new last-seen. Only kind-4 contributes
-          // here — NIP-59 wraps have randomized timestamps (~2 days in
-          // either direction of the real publish time) for plausible
-          // deniability, so wrap.created_at can't be used as a
-          // monotonic publish-time cursor. Including them here would
-          // ratchet lastSeen into the future on the first wrap with a
-          // forward-dated ts, then cause subsequent kind-4 since-filters
-          // to drop legitimate recent NIP-04 messages. fetchInboxDmEvents
-          // already drops the `since` filter for kind-1059 entirely (see
-          // the matching comment there); the cache dedupes wraps by id.
-          const newLastSeen = Math.max(lastSeen ?? 0, ...kind4.map((e) => e.created_at));
-          await Promise.all([
-            AsyncStorage.setItem(inboxCacheKey(refreshForPubkey), JSON.stringify(merged)).catch(
-              () => {},
-            ),
-            newLastSeen > (lastSeen ?? 0)
-              ? AsyncStorage.setItem(inboxLastSeenKey(refreshForPubkey), String(newLastSeen)).catch(
-                  () => {},
-                )
-              : Promise.resolve(),
-          ]);
-        } catch (error) {
-          if (__DEV__) console.warn('[Nostr] refreshDmInbox failed:', error);
-        } finally {
-          setDmInboxLoading(false);
-        }
-      })();
-
-      dmInboxInFlight.current = { promise: task, includeNonFollows };
-      try {
-        await task;
-        dmInboxLastRefreshAt.current = performance.now();
-      } finally {
-        dmInboxInFlight.current = null;
-        const __perfBlockMs = Math.round(performance.now() - __perfBlockStart);
-        // Only surface costly refreshes — sub-200 ms ones aren't
-        // contributors to the multi-second freezes we're hunting.
-        if (__perfBlockMs > 200) {
-          console.log(`[PerfBlock] refreshDmInbox: ${__perfBlockMs}ms`);
-        }
-      }
-    },
-    [
-      pubkey,
-      isLoggedIn,
-      signerType,
-      getReadRelays,
-      followPubkeys,
-      decryptNip04ViaSigner,
-      amberNip44DecryptSilent,
-    ],
-  );
-
-  useEffect(() => {
-    if (!isLoggedIn) setDmInbox([]);
-  }, [isLoggedIn]);
-
-  // Live mirror of `followPubkeys` for the long-lived kind-1059
-  // subscription below. The sub captures `followPubkeys` at the time
-  // the effect ran; without this ref, a follow added after sub
-  // creation would be invisible to the gate until the sub
-  // reconnected. Reading via ref keeps the gate fresh per event
-  // without thrashing the subscription on every contacts update.
-  const followPubkeysRef = useRef(followPubkeys);
-  useEffect(() => {
-    followPubkeysRef.current = followPubkeys;
-  }, [followPubkeys]);
-
-  // Idempotent — any DM-surface (Messages tab, ConversationScreen)
-  // calls this on focus. The first call flips `liveSubArmed`, the
-  // gated useEffect below re-runs and opens the live NIP-17 sub.
-  // Subsequent calls are no-ops (React bails on identical setState).
-  const armLiveDmSub = useCallback(() => {
-    setLiveSubArmed(true);
-  }, []);
-
-  // In-memory dedup Set that survives live-DM-sub re-opens. The sub
-  // useEffect below re-runs when getReadRelays changes — e.g. when the
-  // relay-list refresh adds a new relay 9 s into cold start. Without
-  // this ref, the new effect instance creates a fresh Set and re-seeds
-  // it from AsyncStorage's wrap cache. That snapshot is stale by the
-  // deferred-write window, so all wraps the prior sub already
-  // decrypted re-stream from the relays (same `since` cursor) and get
-  // re-routed/re-decrypted. Carrying the Set forward keeps the
-  // early-return in handleInboxEvent honest across the re-open.
-  // Reset only when the viewer changes (sign out / account switch).
-  const knownWrapIdsRef = useRef<{ pubkey: string | null; set: Set<string> }>({
-    pubkey: null,
-    set: new Set(),
-  });
-
-  // Long-lived kind-1059 (NIP-17 gift wrap) subscription for the
-  // current viewer (#349). Without this, new incoming wraps only
-  // surface via pull-to-refresh or the 30 s-TTL useFocusEffect on
-  // MessagesScreen — which means the user sits on the Messages tab
-  // for up to half a minute after a friend sends a DM with nothing
-  // happening on screen.
-  //
-  // Per-event handler:
-  //  1. Dedupe against (a) a session-scoped `seen` set so the same
-  //     wrap delivered by multiple relays is processed once, and
-  //     (b) the persisted Nip17CacheEntry cache so wraps previously
-  //     decrypted by `refreshDmInbox` short-circuit.
-  //  2. Decrypt with the active signer's NIP-17 helper — same code
-  //     path used by `refreshDmInbox` (`unwrapWrapNsec` for nsec,
-  //     `unwrapWrapViaNip44` + Amber silent-decrypt for Amber).
-  //  3. Try `tryRouteGroupRumor` first. Multi-recipient kind-14
-  //     rumors land in group storage and fire the existing
-  //     `notifyGroupMessage` listener — open GroupConversationScreen
-  //     re-loads automatically.
-  //  4. 1:1 rumors that pass the follow gate are written to the
-  //     persistent NIP-17 wrap cache (so the next inbox / thread open
-  //     can short-circuit), appended to `dmInbox` state, and
-  //     broadcast to `dmMessageListeners` so an open
-  //     ConversationScreen for that peer re-fetches.
-  //
-  // Follow gate: matches `refreshDmInbox`'s default — non-followed
-  // sender wraps are decrypted (so we can group-route them) but NOT
-  // cached or surfaced to dmInbox state. The dev-mode "All (dev)"
-  // toggle still relies on the next pull-to-refresh to surface
-  // unfollowed live wraps; live delivery for that view is a
-  // follow-up. Rationale: caching unfollowed plaintext on disk
-  // violates the "B1 — never cache rumors from non-followed senders"
-  // invariant in `refreshDmInbox`.
-  //
-  // Writes to the wrap + inbox caches go through a serial queue to
-  // avoid racing with `refreshDmInbox` (both read-modify-write the
-  // same AsyncStorage blobs). The queue is per-effect-instance; the
-  // single-flight guard in `refreshDmInbox` serialises on its side.
-  useEffect(() => {
-    if (!isLoggedIn || !pubkey || !signerType) return;
-    // Wait until a DM-surface (Messages tab, ConversationScreen) has
-    // focused at least once before opening the live sub. On cold boot
-    // the user is on Home, so we skip ~5 s of per-wrap unwrap/route/
-    // dedup JS-thread work. First Messages focus flips `liveSubArmed`,
-    // this effect re-runs, sub opens, drain happens then — when the
-    // user is explicitly looking at messages and a brief loading
-    // state is expected.
-    if (!liveSubArmed) return;
-    const viewerPubkey = pubkey;
-    const activeSigner = signerType;
-    const readRelays = getReadRelays();
-    const seen = new Set<string>();
-    const SEEN_CAP = 4096;
-    // In-memory mirror of the persisted NIP-17 wrap-id cache. Backed
-    // by `knownWrapIdsRef` so the Set survives this effect's re-runs
-    // (relay-list change → fresh effect instance). Seeded by union
-    // below from AsyncStorage's wrap cache, but does NOT replace any
-    // entries the prior sub instance added in-memory but the deferred
-    // writeChain hasn't persisted yet. Per issue #505 — the relay
-    // re-streams the backlog since the last `since` cursor, and on a
-    // busy account that's 100+ wraps in ~12 s, almost all of which are
-    // already known; pre-#505 the dedup-cache hit check was
-    // lazy-populated and downstream of several per-event operations.
-    if (knownWrapIdsRef.current.pubkey !== viewerPubkey) {
-      knownWrapIdsRef.current = { pubkey: viewerPubkey, set: new Set() };
-    }
-    const knownWrapIds: Set<string> = knownWrapIdsRef.current.set;
-    let cancelled = false;
-    let writeChain: Promise<void> = Promise.resolve();
-
-    // Coalesce per-event inbox merges into one setDmInbox call per ~150 ms or per 25 events. Without this, a relay-restream burst (e.g. cold start with 200+ kind-4 events queued) causes one React re-render per event = 30+ rerenders/sec on the JS thread, which is what locks the UI for 30 seconds. Batching collapses that into ~6 rerenders/sec at most. Notifications still fire per-event so unread counts/sounds aren't dropped.
-    let pendingInboxEntries: DmInboxEntry[] = [];
-    let pendingFlushTimer: ReturnType<typeof setTimeout> | null = null;
-    const PENDING_FLUSH_MS = 150;
-    const PENDING_FLUSH_THRESHOLD = 25;
-    const flushPendingInbox = (): void => {
-      if (pendingFlushTimer) {
-        clearTimeout(pendingFlushTimer);
-        pendingFlushTimer = null;
-      }
-      if (pendingInboxEntries.length === 0) return;
-      const batch = pendingInboxEntries;
-      pendingInboxEntries = [];
-      setDmInbox((prev) => mergeInboxEntries(prev, batch, DM_INBOX_CAP));
-    };
-    const queueInboxEntry = (entry: DmInboxEntry): void => {
-      pendingInboxEntries.push(entry);
-      if (pendingInboxEntries.length >= PENDING_FLUSH_THRESHOLD) {
-        flushPendingInbox();
-        return;
-      }
-      if (pendingFlushTimer === null) {
-        pendingFlushTimer = setTimeout(flushPendingInbox, PENDING_FLUSH_MS);
-      }
-    };
-
-    const handleInboxEvent = async (ev: nostrService.RawInboxDmEvent): Promise<void> => {
-      // Earliest-possible short-circuit for NIP-17 wraps we already
-      // decrypted on a previous launch. Cost saved per backlog wrap:
-      // console.log + seen.has + seen.add + kind dispatch + cacheKey
-      // build + the async AsyncStorage.getItem race. On a busy
-      // fixture with 100+ wraps in the backlog this compressed the
-      // cold-start JS-thread occupation window from ~12 s to <1 s.
-      // Per issue #505.
-      if (ev.kind === 1059 && knownWrapIds.has(ev.id)) return;
-      // Eagerly claim this wrap.id in the in-memory dedup Set before
-      // doing any async work. The deferred writeChain at the bottom of
-      // this handler also does this, but only after AsyncStorage I/O
-      // completes — leaving a window where a re-opened sub (relay-list
-      // change mid cold start) re-streams the same wrap and gets past
-      // the early-return because the Set hasn't been updated yet.
-      // Set.add is idempotent so the trailing writeChain add becomes
-      // a no-op. kind 4 has its own `seen` Set below.
-      if (ev.kind === 1059) knownWrapIds.add(ev.id);
-      if (__DEV__) console.log(`[Nostr] live evt kind=${ev.kind} recv ${ev.id.slice(0, 8)}`);
-      if (cancelled) return;
-      if (seen.has(ev.id)) {
-        if (__DEV__) console.log(`[Nostr] live evt ${ev.id.slice(0, 8)} dedup-seen`);
-        return;
-      }
-      seen.add(ev.id);
-      // Drop oldest ~25% so a long-lived sub under spam doesn't grow the Set unboundedly.
-      if (seen.size > SEEN_CAP) {
-        const drop = Math.floor(SEEN_CAP / 4);
-        const it = seen.values();
-        for (let i = 0; i < drop; i++) seen.delete(it.next().value!);
-      }
-
-      // NIP-04 (kind-4) — partner is in the envelope; decrypt directly with the active signer.
-      if (ev.kind === 4) {
-        const fromMe = ev.pubkey === viewerPubkey;
-        const recipientTag = ev.tags.find((t) => t[0] === 'p')?.[1]?.toLowerCase();
-        const partnerPubkey = fromMe ? recipientTag : ev.pubkey.toLowerCase();
-        if (!partnerPubkey || !/^[0-9a-f]{64}$/.test(partnerPubkey)) {
-          if (__DEV__) console.log(`[Nostr] live kind-4 ${ev.id.slice(0, 8)} no-partner`);
-          return;
-        }
-        let plaintext = nip04PlaintextCache.get(ev.id);
-        if (plaintext === undefined) {
-          try {
-            if (activeSigner === 'nsec') {
-              const secretKey = await getMemoisedSecretKey(viewerPubkey);
-              if (!secretKey) return;
-              plaintext = await nostrService.decryptNip04WithSecret(
-                secretKey,
-                partnerPubkey,
-                ev.content,
-              );
-            } else if (activeSigner === 'amber') {
-              plaintext = await amberService.requestNip04Decrypt(
-                ev.content,
-                partnerPubkey,
-                viewerPubkey,
-              );
-            } else {
-              return;
-            }
-          } catch (error) {
-            if (__DEV__)
-              console.warn(`[Nostr] live kind-4 ${ev.id.slice(0, 8)} decrypt failed:`, error);
-            return;
-          }
-          if (!plaintext) {
-            if (__DEV__) console.log(`[Nostr] live kind-4 ${ev.id.slice(0, 8)} empty-plaintext`);
-            return;
-          }
-          nip04PlaintextCache.set(ev.id, plaintext);
-        } else if (__DEV__) {
-          console.log(`[Nostr] live kind-4 ${ev.id.slice(0, 8)} dedup-cache`);
-        }
-        if (cancelled || viewerPubkey !== pubkey || activeSigner !== signerType) return;
-
-        // Follow gate (mirrors refreshDmInbox B1) — incoming kind-4 from a non-followed sender is dropped from inbox state. Outgoing (fromMe) bypasses since we sent it.
-        if (!fromMe && !followPubkeysRef.current.has(partnerPubkey)) {
-          if (__DEV__)
-            console.log(
-              `[Nostr] live kind-4 ${ev.id.slice(0, 8)} dropped by follow-gate (partner=${partnerPubkey.slice(0, 8)})`,
-            );
-          return;
-        }
-
-        const k4InboxEntry: DmInboxEntry = {
-          id: ev.id,
-          partnerPubkey,
-          fromMe,
-          createdAt: ev.created_at,
-          text: plaintext,
-          wireKind: 4,
-        };
-        // No wrap-id cache for kind-4 (plaintext lives in RAM-only LRU); only persist the inbox preview blob. Same writeChain as kind-1059 to serialize concurrent inbox writes. Also bump inboxLastSeenKey so refreshDmInbox's kind-4 `since` filter advances and doesn't re-fetch already-seen events on the next refresh.
-        writeChain = writeChain
-          .then(async () => {
-            if (cancelled) return;
-            const inboxRaw = await safeGetDmCacheItem(inboxCacheKey(viewerPubkey));
-            const cachedInbox: DmInboxEntry[] = inboxRaw
-              ? (() => {
-                  try {
-                    const parsed = JSON.parse(inboxRaw);
-                    return Array.isArray(parsed) ? parsed : [];
-                  } catch {
-                    return [];
-                  }
-                })()
-              : [];
-            const merged = mergeInboxEntries(cachedInbox, [k4InboxEntry], DM_INBOX_CAP);
-            // Re-check after the await: logout may have multiRemove'd these keys while we were reading. Without this, a freshly-decrypted DM would re-populate disk after the user signed out.
-            if (cancelled) return;
-            await AsyncStorage.setItem(inboxCacheKey(viewerPubkey), JSON.stringify(merged)).catch(
-              () => {},
-            );
-            const lastSeenRaw = await AsyncStorage.getItem(inboxLastSeenKey(viewerPubkey));
-            const lastSeen = lastSeenRaw ? Number(lastSeenRaw) : 0;
-            if (ev.created_at > lastSeen) {
-              if (cancelled) return;
-              await AsyncStorage.setItem(
-                inboxLastSeenKey(viewerPubkey),
-                String(ev.created_at),
-              ).catch(() => {});
-            }
-          })
-          .catch((e) => {
-            if (__DEV__) console.warn('[Nostr] live kind-4 persist failed:', e);
-          });
-        await writeChain;
-        if (cancelled || viewerPubkey !== pubkey || activeSigner !== signerType) return;
-
-        queueInboxEntry(k4InboxEntry);
-        notifyDmMessage(partnerPubkey);
-        if (__DEV__)
-          console.log(
-            `[Nostr] live kind-4 ${ev.id.slice(0, 8)} surfaced (partner=${partnerPubkey.slice(0, 8)}, fromMe=${fromMe})`,
-          );
-        return;
-      }
-
-      // NIP-17 (kind-1059) — existing gift-wrap unwrap path. Local alias preserves original variable name without renaming through the body below.
-      const wrap = ev;
-
-      // Cache short-circuit: if refreshDmInbox already decrypted this
-      // wrap and persisted it, the live sub has nothing to do — the
-      // event was either delivered before the sub opened or arrived
-      // via two paths (live + a near-simultaneous force-refresh).
-      const cacheKey =
-        activeSigner === 'nsec'
-          ? perAccountKey(NSEC_NIP17_CACHE_KEY_BASE, viewerPubkey)
-          : activeSigner === 'amber'
-            ? perAccountKey(AMBER_NIP17_CACHE_KEY_BASE, viewerPubkey)
-            : null;
-      if (!cacheKey) return;
-      // knownWrapIds is seeded eagerly up-front below (before the
-      // subscription opens) — the in-flow lazy-load was removed in
-      // #505 because it (a) raced when many wraps arrived together
-      // and each tried to seed the Set concurrently, and (b) made
-      // dedup hits pay through a long per-event prologue before the
-      // check fired. The check for cached IDs now lives at the very
-      // top of this function for kind-1059 events. This line is only
-      // reached for genuinely new (not-yet-cached) wraps, OR — in
-      // the rare case the seed failed (see the catch in the sub-open
-      // block) — for wraps that should have been pre-known. In that
-      // case the wrap re-decrypts; the persistent `wrapCache` write
-      // below still guards against the on-disk cache filling
-      // unboundedly, but the `dmMessageListeners` may fire a second
-      // time for messages already shown in a previous session.
-      // Acceptable trade-off because the seed only fails on
-      // AsyncStorage I/O error which is extremely rare on Android.
-
-      const onSkip = (reason: string, wrapId: string) => {
-        if (__DEV__) console.warn(`[Nostr] live NIP-17 unwrap skip (${wrapId}): ${reason}`);
-      };
-
-      // Yield to the event loop before each per-wrap decryption. The
-      // live sub fans out wraps from the relay one at a time, but
-      // when the sub catches up a backlog after cold start, multiple
-      // wraps land in the same JS task — each sync `unwrapWrapNsec`
-      // is ~1-3 ms and they pile up to tens-of-ms of unbroken
-      // blocking, dropping bottom-sheet animation frames. A single
-      // setTimeout(0) per wrap costs ~0 ms but lets RN re-flush
-      // pending UI events between decryptions. See issue #496.
-      await yieldToEventLoop();
-
-      let rumor: DecodedRumor | null = null;
-      if (activeSigner === 'nsec') {
-        const secretKey = await getMemoisedSecretKey(viewerPubkey);
-        if (!secretKey) return;
-        rumor = unwrapWrapNsec(wrap, secretKey, onSkip);
-      } else if (activeSigner === 'amber') {
-        try {
-          rumor = await unwrapWrapViaNip44(
-            wrap,
-            (ct, cp) => amberService.requestNip44DecryptSilent(ct, cp, viewerPubkey),
-            onSkip,
-          );
-        } catch (error) {
-          const code = (error as { code?: string })?.code;
-          const message = (error as Error)?.message ?? '';
-          if (code === 'PERMISSION_NOT_GRANTED' || /PERMISSION_NOT_GRANTED/.test(message)) {
-            // Same flag refreshDmInbox sets — Account screen surfaces
-            // a one-tap grant button; without it, the live sub would
-            // silently fail every wrap until the user re-enabled
-            // Amber's blanket nip44_decrypt.
-            setAmberNip44Permission('denied');
-            return;
-          }
-          if (__DEV__) console.warn('[Nostr] live Amber NIP-17 unwrap failed:', error);
-          return;
-        }
-      }
-      if (!rumor) {
-        if (__DEV__) console.log(`[Nostr] live wrap ${wrap.id.slice(0, 8)} no-rumor`);
-        return;
-      }
-      if (cancelled || viewerPubkey !== pubkey || activeSigner !== signerType) return;
-
-      // Group-route first — multi-recipient rumors are owned by the
-      // group surface, not the 1:1 inbox. tryRouteGroupRumor handles
-      // appendGroupMessage + notifyGroupMessage internally so an open
-      // GroupConversationScreen auto-refreshes.
-      const routeResult = await tryRouteGroupRumor(rumor, viewerPubkey, wrap.id);
-      if (routeResult.kind !== 'not-group') {
-        if (__DEV__)
-          console.log(
-            `[Nostr] live wrap ${wrap.id.slice(0, 8)} group-routed (${routeResult.kind})`,
-          );
-        return;
-      }
-
-      const partnership = partnerFromRumor(rumor, viewerPubkey);
-      if (!partnership) {
-        if (__DEV__) console.log(`[Nostr] live wrap ${wrap.id.slice(0, 8)} no-partnership`);
-        return;
-      }
-
-      // Follow gate (mirrors refreshDmInbox B1) — keeps non-followed
-      // sender plaintext off AsyncStorage. Group rumors above don't
-      // hit this gate because group membership is its own auth signal.
-      if (!followPubkeysRef.current.has(partnership.partnerPubkey)) {
-        if (__DEV__)
-          console.log(
-            `[Nostr] live wrap ${wrap.id.slice(0, 8)} dropped by follow-gate (partner=${partnership.partnerPubkey.slice(0, 8)})`,
-          );
-        return;
-      }
-
-      const entry: Nip17CacheEntry = {
-        id: wrap.id,
-        wrapId: wrap.id,
-        partnerPubkey: partnership.partnerPubkey,
-        fromMe: partnership.fromMe,
-        createdAt: rumor.created_at,
-        text: rumor.content,
-        wireKind: rumor.kind,
-      };
-      const inboxEntry: DmInboxEntry = {
-        id: entry.id,
-        partnerPubkey: entry.partnerPubkey,
-        fromMe: entry.fromMe,
-        createdAt: entry.createdAt,
-        text: entry.text,
-        wireKind: entry.wireKind,
-      };
-
-      // Serialise read→merge→write of wrap+inbox blobs so concurrent live wraps don't race each other.
-      // The trailing `.catch` is load-bearing: without it a single throw leaves `writeChain` rejected and every later `.then(...)` skips its onFulfilled.
-      writeChain = writeChain
-        .then(async () => {
-          if (cancelled) return;
-          const wrapRaw = await AsyncStorage.getItem(cacheKey);
-          const wrapCache = safeParseRecord<Nip17CacheEntry>(wrapRaw);
-          if (wrapCache[wrap.id]) {
-            knownWrapIds?.add(wrap.id);
-            return;
-          }
-          wrapCache[wrap.id] = entry;
-          // Re-check after each await: logout may have multiRemove'd these keys while we were reading. Without these guards a freshly-decrypted wrap would re-populate disk after the user signed out.
-          if (cancelled) return;
-          await writeNip17Cache(cacheKey, wrapCache);
-          knownWrapIds?.add(wrap.id);
-
-          const inboxRaw = await safeGetDmCacheItem(inboxCacheKey(viewerPubkey));
-          const cachedInbox: DmInboxEntry[] = inboxRaw
-            ? (() => {
-                try {
-                  const parsed = JSON.parse(inboxRaw);
-                  return Array.isArray(parsed) ? parsed : [];
-                } catch {
-                  return [];
-                }
-              })()
-            : [];
-          const merged = mergeInboxEntries(cachedInbox, [inboxEntry], DM_INBOX_CAP);
-          if (cancelled) return;
-          await AsyncStorage.setItem(inboxCacheKey(viewerPubkey), JSON.stringify(merged)).catch(
-            () => {},
-          );
-        })
-        .catch((e) => {
-          if (__DEV__) console.warn('[Nostr] live wrap persist failed:', e);
-        });
-      await writeChain;
-      if (cancelled || viewerPubkey !== pubkey || activeSigner !== signerType) return;
-
-      queueInboxEntry(inboxEntry);
-      notifyDmMessage(partnership.partnerPubkey);
-      if (__DEV__)
-        console.log(
-          `[Nostr] live wrap ${wrap.id.slice(0, 8)} surfaced (partner=${partnership.partnerPubkey.slice(0, 8)})`,
-        );
-    };
-
-    // Load the persisted kind-4 lastSeen cursor before opening the sub so the relay only re-streams events the user hasn't seen yet — without this, a heavy DM history floods the JS thread with hundreds of `live evt kind=4` deliveries on every cold start (each one a NIP-04 decrypt round-trip + setDmInbox re-render).
-    let unsubscribe: (() => void) | null = null;
-    (async () => {
-      // Reuse loadLastSeen so parsing/validation matches refreshDmInbox's existing reads of the same key (#409 review). loadLastSeen returns undefined for missing/invalid values, which subscribeInboxDmsForViewer then falls back to its 7-day floor for.
-      const sinceK4 = await loadLastSeen(inboxLastSeenKey(viewerPubkey)).catch(() => undefined);
-      if (cancelled) return;
-      // Pre-seed `knownWrapIds` from the persisted NIP-17 wrap-id
-      // cache. One JSON.parse here saves N inline AsyncStorage reads
-      // + parses inside `handleInboxEvent` when the relay re-streams
-      // the backlog. The early-return in `handleInboxEvent` (top)
-      // checks this Set as the very first thing and skips all
-      // downstream per-event work for cache hits. Per issue #505.
-      const wrapCacheKey =
-        activeSigner === 'nsec'
-          ? perAccountKey(NSEC_NIP17_CACHE_KEY_BASE, viewerPubkey)
-          : activeSigner === 'amber'
-            ? perAccountKey(AMBER_NIP17_CACHE_KEY_BASE, viewerPubkey)
-            : null;
-      if (wrapCacheKey) {
-        try {
-          const seedRaw = await AsyncStorage.getItem(wrapCacheKey);
-          const seedCache = safeParseRecord<Nip17CacheEntry>(seedRaw);
-          for (const id of Object.keys(seedCache)) knownWrapIds.add(id);
-          // Also seed from persisted group messages — group-routed wraps
-          // never land in the 1:1 wrapCache (the tryRouteGroupRumor
-          // branch returns before the cache write). Without this union
-          // every cold start re-decrypts + re-routes the same group
-          // wraps the relay re-streams since the last `since` cursor.
-          const dmCount = knownWrapIds.size;
-          const groupWrapIds = await listPersistedGroupWrapIds();
-          for (const id of groupWrapIds) knownWrapIds.add(id);
-          if (__DEV__) {
-            console.log(
-              `[Nostr] live DM sub: seeded knownWrapIds with ${dmCount} dm + ${knownWrapIds.size - dmCount} group wraps`,
-            );
-          }
-        } catch (e) {
-          // Seed-from-disk failed — leave knownWrapIds as the empty
-          // Set we initialised at outer-scope. Cached wraps re-stream
-          // through the full handler (decrypt, route, write-cache,
-          // queueInboxEntry, notifyDmMessage). Two observable side
-          // effects vs the pre-#505 in-flow dedup check:
-          //   1. `dmMessageListeners` registered for an open
-          //      conversation will re-fire for messages already
-          //      surfaced in a prior session.
-          //   2. The `unwrapWrapNsec` / `unwrapWrapViaNip44` call
-          //      runs unnecessarily for each cached wrap (1–3 ms each).
-          // The persistent on-disk wrapCache write is idempotent —
-          // it doesn't grow unboundedly. We accept this regression on
-          // the failure path because (a) AsyncStorage.getItem I/O
-          // errors are extremely rare on Android, and (b) the
-          // alternative (resurrecting the lazy-load inside the
-          // handler) would re-introduce the race + per-event prologue
-          // cost that motivated #505 in the first place.
-          if (__DEV__)
-            console.warn('[Nostr] live DM sub: knownWrapIds seed failed, dedup degraded:', e);
-        }
-      }
-      if (cancelled) return;
-      unsubscribe = nostrService.subscribeInboxDmsForViewer({
-        viewerPubkey,
-        relays: readRelays,
-        sinceK4,
-        onEvent: (ev) => {
-          // Fire-and-forget: handleInboxEvent awaits its own state, and any throw is caught + logged here so the sub keeps running.
-          handleInboxEvent(ev).catch((e) => {
-            if (__DEV__) console.warn('[Nostr] live DM handler failed:', e);
-          });
-        },
-      });
-      if (__DEV__) {
-        console.log(
-          `[Nostr] live DM sub (kinds 4 + 1059) opened for ${viewerPubkey.slice(0, 8)} on ${readRelays.length} relays, sinceK4=${sinceK4 ?? 'default-90d'}`,
-        );
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-      flushPendingInbox();
-      if (unsubscribe) unsubscribe();
-    };
-  }, [isLoggedIn, pubkey, signerType, getReadRelays, liveSubArmed]);
 
   const signEvent = useCallback(
     async (event: {
